AudioTrackSource, on_captured_frame

+++ livekit-webrtc/src/audio_frame.rs
@@ -2,6 +2,6 @@
 pub struct AudioFrame {
     pub data: Vec<i16>,
     pub sample_rate_hz: u32,
-    pub num_channels: u32,
-    pub samples_per_channel: u32,
+    pub num_channels: usize,
+    pub samples_per_channel: usize,
 }
